feat: add ui page
