+++ async-race/src/index.css
@@ -287,6 +287,7 @@ body {
 button:disabled {
   cursor: default;
   background-color: #666666;
+  color: #ffffff;
   box-shadow: none;
 }
 button:disabled:hover {
